requirements.txt
fastapi==0.116.1
firebase-admin==6.3.0
google-cloud-aiplatform==1.104.0
Pillow==11.3.0
protobuf==6.31.1
pydantic==2.11.7
pydub==0.25.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
uvicorn==0.35.0
